requirements.txt
cffi
cryptography
pycparser
PyMySQL==1.0.2
python-dotenv
six
